+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -1532,6 +1532,7 @@ public class BetonQuest extends JavaPlugin {
      * @return the value of this variable for given player
      */
     public String getVariableValue(final String packName, final String name, final String playerID) {
+
         if (!Config.getPackages().containsKey(packName)) {
             log.warn("Variable '" + name + "' contains the non-existent package '" + packName + "' !");
             return "";
@@ -1541,6 +1542,10 @@ public class BetonQuest extends JavaPlugin {
             final Variable var = createVariable(pack, name);
             if (var == null) {
                 log.warn(pack, "Could not resolve variable '" + name + "'.");
+                return "";
+            }
+            if ((playerID == null || "".equals(playerID.trim())) && !var.isStaticness()) {
+                log.warn(pack, "Variable '" + name + "' cannot be executed without a player reference!");
                 return "";
             }
             return var.getValue(playerID);