+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/CitizensHologram.java
@@ -1,8 +1,10 @@
 package org.betonquest.betonquest.compatibility.citizens;
 
-import com.gmail.filoghost.holographicdisplays.api.Hologram;
-import com.gmail.filoghost.holographicdisplays.api.HologramsAPI;
 import lombok.CustomLog;
+import me.filoghost.holographicdisplays.api.HolographicDisplaysAPI;
+import me.filoghost.holographicdisplays.api.hologram.Hologram;
+import me.filoghost.holographicdisplays.api.hologram.PlaceholderSetting;
+import me.filoghost.holographicdisplays.api.hologram.VisibilitySettings;
 import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.npc.NPC;
 import org.betonquest.betonquest.BetonQuest;
@@ -86,7 +88,7 @@ public class CitizensHologram extends BukkitRunnable {
         for (final List<NPCHologram> holograms : npcs.values()) {
             for (final NPCHologram npcHologram : holograms) {
                 if (npcHologram.hologram != null) {
-                    npcHologram.hologram.getVisibilityManager().resetVisibilityAll();
+                    npcHologram.hologram.getVisibilitySettings().clearIndividualVisibilities();
                     npcHologram.hologram.delete();
                     npcHologram.hologram = null;
                 }
@@ -169,7 +171,7 @@ public class CitizensHologram extends BukkitRunnable {
                 if (updateHologramsForPlayers(npcHologram, npc)) {
                     npcUpdater = true;
                 } else if (npcHologram.hologram != null) {
-                    npcHologram.hologram.getVisibilityManager().resetVisibilityAll();
+                    npcHologram.hologram.getVisibilitySettings().clearIndividualVisibilities();
                     npcHologram.hologram.delete();
                     npcHologram.hologram = null;
                 }
@@ -199,7 +201,7 @@ public class CitizensHologram extends BukkitRunnable {
                 if (npcHologram.hologram != null) {
                     final NPC npc = CitizensAPI.getNPCRegistry().getById(entry.getKey());
                     if (npc != null) {
-                        npcHologram.hologram.teleport(npc.getStoredLocation().add(npcHologram.vector));
+                        npcHologram.hologram.setPosition(npc.getStoredLocation().add(npcHologram.vector));
                     }
                 }
             }
@@ -212,17 +214,18 @@ public class CitizensHologram extends BukkitRunnable {
             if (BetonQuest.conditions(PlayerConverter.getID(player), npcHologram.conditions)) {
                 hologramEnabled = true;
                 if (npcHologram.hologram == null) {
-                    final Hologram hologram = HologramsAPI.createHologram(BetonQuest.getInstance(), npc.getStoredLocation().add(npcHologram.vector));
-                    hologram.getVisibilityManager().setVisibleByDefault(false);
+                    final Hologram hologram = HolographicDisplaysAPI.get(BetonQuest.getInstance()).createHologram(npc.getStoredLocation().add(npcHologram.vector));
+                    hologram.setPlaceholderSetting(PlaceholderSetting.ENABLE_ALL);
+                    hologram.getVisibilitySettings().setGlobalVisibility(VisibilitySettings.Visibility.HIDDEN);
                     updateHologramForPlayersLines(npcHologram, hologram);
                     npcHologram.hologram = hologram;
                 }
-                if (!npcHologram.hologram.getVisibilityManager().isVisibleTo(player)) {
-                    npcHologram.hologram.getVisibilityManager().showTo(player);
+                if (!npcHologram.hologram.getVisibilitySettings().isVisibleTo(player)) {
+                    npcHologram.hologram.getVisibilitySettings().setIndividualVisibility(player, VisibilitySettings.Visibility.VISIBLE);
                 }
             } else {
-                if (npcHologram.hologram != null && npcHologram.hologram.getVisibilityManager().isVisibleTo(player)) {
-                    npcHologram.hologram.getVisibilityManager().hideTo(player);
+                if (npcHologram.hologram != null && npcHologram.hologram.getVisibilitySettings().isVisibleTo(player)) {
+                    npcHologram.hologram.getVisibilitySettings().setIndividualVisibility(player, VisibilitySettings.Visibility.HIDDEN);
                 }
             }
         }
@@ -242,14 +245,14 @@ public class CitizensHologram extends BukkitRunnable {
                         stackSize = 1;
                     }
                     final ItemStack stack = new QuestItem(itemID).generate(stackSize);
-                    hologram.appendItemLine(stack);
+                    hologram.getLines().appendItem(stack);
                 } catch (final InstructionParseException e) {
                     LOG.warn(npcHologram.pack, "Could not parse item in " + npcHologram.pack.getPackagePath() + " hologram: " + e.getMessage(), e);
                 } catch (final ObjectNotFoundException e) {
                     LOG.warn(npcHologram.pack, "Could not find item in " + npcHologram.pack.getPackagePath() + " hologram: " + e.getMessage(), e);
                 }
             } else {
-                hologram.appendTextLine(line.replace('&', '§'));
+                hologram.getLines().appendText(line.replace('&', '§'));
             }
         }
     }