+++ src/main/java/org/betonquest/betonquest/variables/GlobalTagVariable.java
@@ -18,6 +18,7 @@ public class GlobalTagVariable extends TagVariable {
      */
     public GlobalTagVariable(final Instruction instruction) throws InstructionParseException {
         super(instruction);
+        staticness = true;
     }
 
     @Override